Extract common admin code into common admin file
