+++ aidants_connect/admin.py
@@ -1,0 +1,86 @@
+from admin_honeypot.admin import LoginAttemptAdmin as HoneypotLoginAttemptAdmin
+from admin_honeypot.models import LoginAttempt as HoneypotLoginAttempt
+from django.contrib.admin import SimpleListFilter
+from django_celery_beat.admin import (
+    ClockedScheduleAdmin,
+    PeriodicTaskAdmin,
+)
+from django_celery_beat.models import (
+    ClockedSchedule,
+    CrontabSchedule,
+    IntervalSchedule,
+    PeriodicTask,
+    SolarSchedule,
+)
+from django_otp.admin import OTPAdminSite
+from magicauth.models import MagicToken
+from aidants_connect_web.models import (
+    DatavizRegion,
+)
+
+admin_site = OTPAdminSite(OTPAdminSite.name)
+admin_site.login_template = "aidants_connect_web/admin/login.html"
+
+admin_site.register(HoneypotLoginAttempt, HoneypotLoginAttemptAdmin)
+
+
+class VisibleToAdminMetier:
+    """A mixin to make a model registered in the Admin visible to Admin Métier.
+    Admin Métier corresponds to the traditional django `is_staff`
+    """
+
+    def has_module_permission(self, request):
+        return request.user.is_staff or request.user.is_superuser
+
+    def has_view_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+    def has_add_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+    def has_change_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+    def has_delete_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+
+class VisibleToTechAdmin:
+    """A mixin to make a model registered in the Admin visible to Tech Admins.
+    ATAC is modelised by is_superuser
+    """
+
+    def has_module_permission(self, request):
+        return request.user.is_superuser
+
+    def has_view_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+    def has_add_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+    def has_change_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+    def has_delete_permission(self, request, obj=None):
+        return self.has_module_permission(request)
+
+
+class RegionFilter(SimpleListFilter):
+    title = "Région"
+
+    parameter_name = "region"
+
+    def lookups(self, request, model_admin):
+        return [(r.id, r.name) for r in DatavizRegion.objects.all()] + [
+            ("other", "Autre")
+        ]
+
+
+# Also register the Django Celery Beat models
+admin_site.register(PeriodicTask, PeriodicTaskAdmin)
+admin_site.register(IntervalSchedule)
+admin_site.register(CrontabSchedule)
+admin_site.register(SolarSchedule)
+admin_site.register(ClockedSchedule, ClockedScheduleAdmin)
+admin_site.register(MagicToken)
