+++ aidants_connect_web/admin.py
@@ -3,8 +3,6 @@ import operator
 from collections.abc import Collection
 from functools import reduce
 
-from admin_honeypot.admin import LoginAttemptAdmin as HoneypotLoginAttemptAdmin
-from admin_honeypot.models import LoginAttempt as HoneypotLoginAttempt
 from django.contrib import messages
 from django.contrib.admin import ModelAdmin, TabularInline, SimpleListFilter
 from django.contrib.auth.admin import UserAdmin as DjangoUserAdmin
@@ -14,18 +12,6 @@ from django.shortcuts import render
 from django.urls import reverse, path
 from django.utils.html import format_html_join
 from django.utils.safestring import mark_safe
-from django_celery_beat.admin import (
-    ClockedScheduleAdmin,
-    PeriodicTaskAdmin,
-)
-from django_celery_beat.models import (
-    ClockedSchedule,
-    CrontabSchedule,
-    IntervalSchedule,
-    PeriodicTask,
-    SolarSchedule,
-)
-from django_otp.admin import OTPAdminSite
 from django_otp.plugins.otp_static.admin import StaticDeviceAdmin
 from django_otp.plugins.otp_static.lib import add_static_token
 from django_otp.plugins.otp_static.models import StaticDevice
@@ -36,10 +22,15 @@ from import_export.admin import ImportMixin, ExportMixin, ImportExportMixin
 from import_export.fields import Field
 from import_export.results import RowResult
 from import_export.widgets import ManyToManyWidget, ForeignKeyWidget
-from magicauth.models import MagicToken
 from nested_admin import NestedModelAdmin, NestedTabularInline
 from tabbed_admin import TabbedModelAdmin
 
+from aidants_connect.admin import (
+    admin_site,
+    VisibleToAdminMetier,
+    VisibleToTechAdmin,
+    RegionFilter,
+)
 from aidants_connect_web.forms import (
     AidantChangeForm,
     AidantCreationForm,
@@ -59,65 +50,7 @@ from aidants_connect_web.models import (
     DatavizDepartmentsToRegion,
 )
 
-admin_site = OTPAdminSite(OTPAdminSite.name)
-admin_site.login_template = "aidants_connect_web/admin/login.html"
-
-admin_site.register(HoneypotLoginAttempt, HoneypotLoginAttemptAdmin)
-
 logger = logging.getLogger()
-
-
-class VisibleToAdminMetier:
-    """A mixin to make a model registered in the Admin visible to Admin Métier.
-    Admin Métier corresponds to the traditional django `is_staff`
-    """
-
-    def has_module_permission(self, request):
-        return request.user.is_staff or request.user.is_superuser
-
-    def has_view_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-    def has_add_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-    def has_change_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-    def has_delete_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-
-class VisibleToTechAdmin:
-    """A mixin to make a model registered in the Admin visible to Tech Admins.
-    ATAC is modelised by is_superuser
-    """
-
-    def has_module_permission(self, request):
-        return request.user.is_superuser
-
-    def has_view_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-    def has_add_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-    def has_change_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-    def has_delete_permission(self, request, obj=None):
-        return self.has_module_permission(request)
-
-
-class RegionFilter(SimpleListFilter):
-    title = "Région"
-
-    parameter_name = "region"
-
-    def lookups(self, request, model_admin):
-        return [(r.id, r.name) for r in DatavizRegion.objects.all()] + [
-            ("other", "Autre")
-        ]
 
 
 class StaticDeviceStaffAdmin(VisibleToAdminMetier, StaticDeviceAdmin):
@@ -1343,14 +1276,6 @@ admin_site.register(Mandat, MandatAdmin)
 admin_site.register(Journal, JournalAdmin)
 admin_site.register(Connection, ConnectionAdmin)
 
-admin_site.register(MagicToken)
 admin_site.register(StaticDevice, StaticDeviceStaffAdmin)
 admin_site.register(TOTPDevice, TOTPDeviceStaffAdmin)
 admin_site.register(CarteTOTP, CarteTOTPAdmin)
-
-# Also register the Django Celery Beat models
-admin_site.register(PeriodicTask, PeriodicTaskAdmin)
-admin_site.register(IntervalSchedule)
-admin_site.register(CrontabSchedule)
-admin_site.register(SolarSchedule)
-admin_site.register(ClockedSchedule, ClockedScheduleAdmin)
